requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.17.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
wandb>=0.15.0
pillow>=9.5.0
safetensors>=0.3.1
huggingface_hub>=0.15.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.2.2
matplotlib>=3.7.1
opencv-python>=4.7.0
torchmetrics>=1.0.0
peft>=0.5.0 